requirements.txt
Django~=4.2.0
djangorestframework~=3.15
requests~=2.26.0
psycopg2-binary~=2.9.1
googlemaps~=4.5.3
pytest~=6.2.5
pytest-django~=4.4.0
requests-mock~=1.9.3
